requirement.txt
pandas==0.23.4
numpy==1.15.4
matplotlib==3.0.2
pymysql==0.9.2
requests==2.21.0
scipy==1.2.1
shapely==1.6.4
py-xgboost
pip install sklearn

